SendMoney controller, fix start up errors, stub handlers/gas calculators/validators

+++ src/main/java/com/boyarsky/dapos/core/tx/GasData.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GasData {
-    private long used;
-    private long wanted;
+    private int used;
+    private int wanted;
 }
