+++ src/main/java/com/boyarsky/dapos/core/tx/type/fee/GasCalculator.java
@@ -4,5 +4,5 @@ import com.boyarsky.dapos.core.tx.Transaction;
 import com.boyarsky.dapos.core.tx.type.TypedComponent;
 
 public interface GasCalculator extends TypedComponent {
-    long gasRequired(Transaction tx);
+    int gasRequired(Transaction tx);
 }
